FIX size display for bruker files
